Fix a bug with misnamed thing
It's super bizarre that it still worked before. Apparently 'event',
which I used as the old argument name for this function, has the same
name as the global variable for the current handled event. Or perhaps
arguments are dynamically binded in javascript?? Anyways, by a second
coincidence, mousemove and mouseup events /both/ have the fields 'x' and
'y', the same fields that I expect inside of this function. Furthermore,
the 'x' and 'y' in the event are the same as the 'x' and 'y' that I
expected in the function in the first place!